% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi_function.R
\name{calc_ndvi}
\alias{calc_ndvi}
\title{Function to calculate NDVI}
\usage{
calc_ndvi(NIR, R)
}
\arguments{
\item{NIR}{near infra-red satellite band (µm)}

\item{R}{Red satellite band (µm)}
}
\value{
NDVI (µm)
}
\description{
Function to calculate NDVI
}
