% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish.R
\docType{data}
\name{fish}
\alias{fish}
\title{Fish catch data}
\format{
A dataframe with 300 observations and 1 variable
itemize{
\item data character string
\item fish_types
}
}
\usage{
fish
}
\description{
Data from ESM 262 Computing for Environmental Science course. Including names of fish caught.
}
\keyword{datasets}
