% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_function.R
\name{fish_calc}
\alias{fish_calc}
\title{fish_calc()}
\usage{
fish_calc(vec)
}
\arguments{
\item{vec}{input vector of items}
}
\value{
common Most common fish

rare the rarest fish

total the total number of fish
}
\description{
Function that takes vector of fish names and always returns three items
}
