% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_stress.R
\name{growth}
\alias{growth}
\title{Function to calculate the effect of heat stress in trees}
\usage{
growth(date, tree_age, ppt, max_temp)
}
\arguments{
\item{date}{YYYY-MM-DD}

\item{tree_age}{age class of tree, 1 = seedling, 2 = sapling, 3 = adult}

\item{ppt}{(mm) total precipitation for the year}

\item{max_temp}{(degrees F) maximum temperature recorded}
}
\value{
text with the risk associated with each temperature and ppt for each age class
}
\description{
Function to calculate the effect of heat stress in trees
}
