% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_function.R
\name{fish}
\alias{fish}
\title{Function to Write a function that takes a vector of fish names and always returns three items}
\usage{
fish(vec)
}
\arguments{
\item{vec}{input vector of items}
}
\value{
common Most common fish

rare the rarest fish

total the total number of fish
}
\description{
Function to Write a function that takes a vector of fish names and always returns three items
}
